cleanup: move RenderFlags to own file, rename Render to RenderCommand
