+++ cmd/abc/abc.go
@@ -37,7 +37,7 @@ var rootCmd = func() *cli.RootCommand {
 					Description: "subcommands for rendering templates and related things",
 					Commands: map[string]cli.CommandFactory{
 						"render": func() cli.Command {
-							return &commands.Render{}
+							return &commands.RenderCommand{}
 						},
 					},
 				}
